feat(classify): add section cards to new-run page and thinking toggle
- ClassificationConfig simplified to labels + classifier type only;
  LLM config (PromptConfigEditor + batch settings) lifted to page level
- NewClassificationRunPage restructured into three Card sections:
  Parse configuration, Classification, LLM configuration
- PromptConfigEditor in classify now passes capabilities={{ thinking: true }}
  to match extract, making the reasoning toggle visible
- ClassificationConfig tests updated to match simplified emitted shape
