+++ frontend/src/components/classification/ClassificationConfig.tsx
@@ -1,5 +1,5 @@
 import { useState } from 'react'
-import { X, ChevronDown } from 'lucide-react'
+import { X } from 'lucide-react'
 import { Button } from '@/components/ui/button'
 import { Input } from '@/components/ui/input'
 import { Label } from '@/components/ui/label'
@@ -11,30 +11,15 @@ import {
   SelectTrigger,
   SelectValue,
 } from '@/components/ui/select'
-import {
-  Collapsible,
-  CollapsibleContent,
-  CollapsibleTrigger,
-} from '@/components/ui/collapsible'
-import { PromptConfigEditor } from '@/components/shared/PromptConfigEditor'
-import type { PromptConfig } from '@/types/prompt-config'
 
 const CLASSIFIER_TYPES = [
   { value: 'llm', label: 'LLM classifier' },
   { value: 'llamaindex_split', label: 'LlamaIndex split (not yet implemented)' },
 ]
 
-const DEFAULT_PROMPT_CONFIG: PromptConfig = {
-  provider: 'ollama_local',
-  model: 'qwen2.5:7b',
-  temperature: 0.0,
-  maxTokens: 4096,
-}
-
 export interface ClassificationConfigValue {
   labels: string[]
   classifierType: string
-  classifierConfig: Record<string, unknown>
 }
 
 interface Props {
@@ -42,66 +27,14 @@ interface Props {
   onChange: (value: ClassificationConfigValue) => void
 }
 
-function configToPromptConfig(config: Record<string, unknown>): PromptConfig {
-  const llm = (config.llm_config as Record<string, unknown> | undefined) ?? {}
-  return {
-    provider: (config.provider as string | undefined) ?? DEFAULT_PROMPT_CONFIG.provider,
-    model: (config.model as string | undefined) ?? DEFAULT_PROMPT_CONFIG.model,
-    temperature: (llm.temperature as number | undefined) ?? DEFAULT_PROMPT_CONFIG.temperature,
-    maxTokens: (llm.max_tokens as number | undefined) ?? DEFAULT_PROMPT_CONFIG.maxTokens,
-    systemPrompt: llm.system_prompt as string | undefined,
-  }
-}
-
-function buildClassifierConfig(
-  classifierType: string,
-  promptConfig: PromptConfig,
-  batchSize: number,
-  batchOverlap: number,
-): Record<string, unknown> {
-  if (classifierType !== 'llm') return {}
-  return {
-    provider: promptConfig.provider,
-    model: promptConfig.model,
-    batch_size: batchSize,
-    batch_overlap: batchOverlap,
-    llm_config: {
-      system_prompt: promptConfig.systemPrompt ?? null,
-      temperature: promptConfig.temperature ?? 0.0,
-      max_tokens: promptConfig.maxTokens ?? 4096,
-    },
-  }
-}
-
 export function ClassificationConfig({ defaultValues, onChange }: Props) {
   const dv = defaultValues ?? {}
   const [labels, setLabels] = useState<string[]>(dv.labels ?? [])
   const [labelInput, setLabelInput] = useState('')
   const [classifierType, setClassifierType] = useState(dv.classifierType ?? 'llm')
-  const [promptConfig, setPromptConfig] = useState<PromptConfig>(
-    dv.classifierConfig && Object.keys(dv.classifierConfig).length > 0
-      ? configToPromptConfig(dv.classifierConfig)
-      : DEFAULT_PROMPT_CONFIG,
-  )
-  const [batchSize, setBatchSize] = useState(
-    (dv.classifierConfig?.batch_size as number | undefined) ?? 10,
-  )
-  const [batchOverlap, setBatchOverlap] = useState(
-    (dv.classifierConfig?.batch_overlap as number | undefined) ?? 3,
-  )
 
-  function emit(
-    nextLabels: string[],
-    nextType: string,
-    nextPrompt: PromptConfig,
-    nextBatch: number,
-    nextOverlap: number,
-  ) {
-    onChange({
-      labels: nextLabels,
-      classifierType: nextType,
-      classifierConfig: buildClassifierConfig(nextType, nextPrompt, nextBatch, nextOverlap),
-    })
+  function emit(nextLabels: string[], nextType: string) {
+    onChange({ labels: nextLabels, classifierType: nextType })
   }
 
   const addLabel = () => {
@@ -110,38 +43,22 @@ export function ClassificationConfig({ defaultValues, onChange }: Props) {
     const next = [...labels, trimmed]
     setLabels(next)
     setLabelInput('')
-    emit(next, classifierType, promptConfig, batchSize, batchOverlap)
+    emit(next, classifierType)
   }
 
   const removeLabel = (l: string) => {
     const next = labels.filter((x) => x !== l)
     setLabels(next)
-    emit(next, classifierType, promptConfig, batchSize, batchOverlap)
+    emit(next, classifierType)
   }
 
   const handleClassifierTypeChange = (v: string) => {
     setClassifierType(v)
-    emit(labels, v, promptConfig, batchSize, batchOverlap)
-  }
-
-  const handlePromptConfigChange = (v: PromptConfig) => {
-    setPromptConfig(v)
-    emit(labels, classifierType, v, batchSize, batchOverlap)
-  }
-
-  const handleBatchSizeChange = (v: number) => {
-    setBatchSize(v)
-    emit(labels, classifierType, promptConfig, v, batchOverlap)
-  }
-
-  const handleBatchOverlapChange = (v: number) => {
-    setBatchOverlap(v)
-    emit(labels, classifierType, promptConfig, batchSize, v)
+    emit(labels, v)
   }
 
   return (
     <div className="space-y-6">
-      {/* Labels */}
       <div className="space-y-2">
         <Label>Labels to classify</Label>
         <div className="flex gap-2">
@@ -178,7 +95,6 @@ export function ClassificationConfig({ defaultValues, onChange }: Props) {
         )}
       </div>
 
-      {/* Classifier type */}
       <div className="space-y-2">
         <Label>Classifier</Label>
         <Select value={classifierType} onValueChange={handleClassifierTypeChange}>
@@ -194,46 +110,6 @@ export function ClassificationConfig({ defaultValues, onChange }: Props) {
           </SelectContent>
         </Select>
       </div>
-
-      {/* LLM config */}
-      {classifierType === 'llm' && (
-        <>
-          <PromptConfigEditor value={promptConfig} onChange={handlePromptConfigChange} />
-          <Collapsible>
-            <CollapsibleTrigger className="flex items-center gap-2 text-sm text-muted-foreground hover:text-foreground">
-              <ChevronDown className="h-4 w-4" />
-              Batch settings
-            </CollapsibleTrigger>
-            <CollapsibleContent className="mt-3 grid grid-cols-2 gap-4">
-              <div className="space-y-2">
-                <Label>Batch size (pages)</Label>
-                <Input
-                  type="number"
-                  min={1}
-                  value={batchSize}
-                  onChange={(e) => handleBatchSizeChange(Number(e.target.value))}
-                />
-              </div>
-              <div className="space-y-2">
-                <Label>Batch overlap (pages)</Label>
-                <Input
-                  type="number"
-                  min={0}
-                  value={batchOverlap}
-                  onChange={(e) => handleBatchOverlapChange(Number(e.target.value))}
-                />
-              </div>
-            </CollapsibleContent>
-          </Collapsible>
-        </>
-      )}
-
-      {/* LlamaIndex placeholder */}
-      {classifierType === 'llamaindex_split' && (
-        <p className="text-sm text-muted-foreground">
-          LlamaIndex split classifier is not yet implemented. Select LLM classifier to proceed.
-        </p>
-      )}
     </div>
   )
 }
